+++ test/helpers.rb
@@ -29,19 +29,21 @@ def flush_frames(count)
   count.times { flush_frame }
 end
 
-def raw_colour_data(data)
-  data.each_slice(10).each { |row|
+def print_raw_colour_data(data, width: 10)
+  puts data.each_slice(width).each { |row|
     puts row.map { |colour|
       "Colour.new(#{colour.red}, #{colour.green}, #{colour.blue}, #{colour.alpha}), "
     }.join
   }
 end
 
-def print_colour_data(data)
+def print_colour_data(data, width: 10)
   map = ""
-  data.each_slice(10).each { |row|
+  data.each_slice(width).each { |row|
     row.each { |colour|
       map << case colour
+      when BLANK
+        "t"
       when RAYWHITE
         "w"
       when BLACK
@@ -65,5 +67,9 @@ def print_colour_data(data)
     map << "\n"
   }
 
-  map
+  puts map
+end
+
+def print_image_data(image)
+  puts print_colour_data(image.data, width: image.width)
 end